ToDo/requirements.txt
asgiref==3.4.1
bookmarks2markdown==1.0.2
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
lxml==4.6.3
Markdown==3.3.4
numpy==1.21.3
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
